# api/src/main/resources/api.yml
openapi: 3.0.3
info:
  title: People Finder
  description: This is the API for the our People Finder application
  version: 1.0.0
  contact:
    email: our-fancy-team@zuehlke.com
paths:
  /people:
    get:
      tags:
        - people
      operationId: getPeople
      summary: Get all people
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Person'
  /skills:
    get:
      tags:
        - skills
      operationId: getSkills
      summary: Get all skills
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Skill'
  /teams:
    post:
      tags:
        - teams
      operationId: calculateTeams
      summary: Calculate all teams
      parameters:
        - in: query
          name: dimension
          example: 3
          schema:
            type: integer
          description: The number of members for the team
      requestBody:
        description: Our team
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Team"
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Team'
components:
  schemas:
    Person:
      properties:
        code:
          type: string
          description: The abbreviation
          example: asdf
          minLength: 3
          maxLength: 4
        firstName:
          type: string
          description: The first name
          example: John
          maxLength: 64
        lastName:
          type: string
          description: The last name
          example: Watson
          maxLength: 64
      required:
        - firstName
        - lastName
    Skill:
      properties:
        name:
          type: string
          description: The skill name
          example: Angular
          maxLength: 64
      required:
        - name
    Proficiency:
      properties:
        rating:
          type: integer
          description: The rating
          minimum: 1
          maximum: 5
        skill:
          $ref: '#/components/schemas/Skill'
      required:
        - rating
        - skill
    Criteria:
      properties:
        proficiencies:
          type: array
          items:
            $ref: '#/components/schemas/Proficiency'
      required:
        - proficiencies
    Team:
      properties:
        name:
          type: string
          description: The team name
          example: Fancy team name
          maxLength: 64
        members:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Criteria'